backend/requirements.txt
fastapi==0.115.12
fastapi-cli==0.0.7
uvicorn==0.34.2
SQLAlchemy==2.0.40
asyncpg==0.30.0
pydantic==2.11.4
pydantic-settings==2.8.1
loguru==0.7.3
python-dotenv==1.1.0
aiosqlite==0.21.0
